+++ grid2op/gym_compat/utils.py
@@ -25,8 +25,17 @@ ATTR_DISCRETE = ("set_line_status",
                  "one_sub_set",
                  "one_sub_change")
 
+
 def check_gym_version():
     import gym
     if gym.__version__ < _MIN_GYM_VERSION:
         raise RuntimeError(f"Grid2op does not work with gym < {_MIN_GYM_VERSION} and you have gym with "
                            f"version {gym.__version__} installed.")
+
+
+def _compute_extra_power_for_losses(gridobj):
+    """
+    to handle the "because of the power losses gen_pmin and gen_pmax can be slightly altered"
+    """
+    import numpy as np
+    return 0.01 * np.sum(np.abs(gridobj.gen_pmax))