+++ CHANGELOG.rst
@@ -24,6 +24,9 @@ Change Log
 
 [1.5.2] - 2021-xx-yy
 -----------------------
+- [FIXED]: `Issue#185 <https://github.com/rte-france/Grid2Op/issues/196>`_ an issue related to the
+  low / high of the observation if using the gym_compat module. Some more protections
+  are enforced now.
 - [IMPROVED]: on windows at least, grid2op does not work with gym < 0.17.2 Checks are performed in order to make sure
   the installed open ai gym package meets this requirement (see issue
   `Issue#185 <https://github.com/rte-france/Grid2Op/issues/185>`_ )